app/requirements.txt
Flask==2.3.3
prometheus_client==0.17.1
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation-flask==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-semantic-conventions==0.43b0
requests==2.32.3